# docs/architecture/reference/openapi.yaml
...
openapi: 3.1.0
info:
  title: OnPrem AI Gateway HTTP API
  version: 0.1.0
  description: >
    HTTP surface of the OnPrem AI Gateway: OpenAI/Anthropic-compatible
    inference endpoints, the portal application API, system administration,
    and the ServerAgent ingest API. Error responses use the stable-code
    envelope `{"error":{"code","message","request_id?"}}`. Session-based
    non-GET/HEAD/OPTIONS requests additionally require the `X-OP-CSRF: 1`
    header. See docs/architecture/reference/api-surface.md for the canonical
    description of auth modes, scopes, and object-level authorization.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: http://127.0.0.1:8080
    description: Local development gateway (make run-gateway / make dev)
  - url: /
    description: >
      Bundled deployment — nginx proxies /api/, /v1/, /openai/, /anthropic/,
      /healthz, and the ACME challenge path to the backend and serves the SPA
      at /portal/.
...
security: []
paths:
...
  /api/portal/usage/groups:
    get:
      tags: [portal]
      summary: Usage grouped (by model/server/…)
      security: [{ bearerAuth: [] }, { sessionCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/OpaqueJSON' }
...
  /api/portal/groups:
    description: >
      User-group collection. Per-group sub-paths (see api-surface.md):
      members, managers, candidates, accept, decline.
    get:
      tags: [portal]
      summary: List user groups
      security: [{ bearerAuth: [] }, { sessionCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/OpaqueJSON' }
    post:
      tags: [portal]
      summary: Create a user group
      security: [{ sessionCookie: [], csrfHeader: [] }, { bearerAuth: [] }]
      requestBody: { $ref: '#/components/requestBodies/OpaqueJSON' }
      responses:
        '200': { $ref: '#/components/responses/OpaqueJSON' }
...
  /api/portal/groups/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [portal]
      summary: Group detail
      security: [{ bearerAuth: [] }, { sessionCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/OpaqueJSON' }
    put:
      tags: [portal]
      summary: Update a group
      security: [{ sessionCookie: [], csrfHeader: [] }, { bearerAuth: [] }]
      requestBody: { $ref: '#/components/requestBodies/OpaqueJSON' }
      responses:
        '200': { $ref: '#/components/responses/OpaqueJSON' }
    delete:
      tags: [portal]
      summary: Delete a group
      security: [{ sessionCookie: [], csrfHeader: [] }, { bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/OpaqueJSON' }
...
  /api/system/netbird/groups:
    get:
      tags: [system]
      summary: NetBird group inspection
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/OpaqueJSON' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: op_ai_gateway_session
      description: >
        Browser session. Non-GET/HEAD/OPTIONS requests additionally require
        the X-OP-CSRF header (modeled as csrfHeader).
    csrfHeader:
      type: apiKey
      in: header
      name: X-OP-CSRF
      description: 'CSRF guard header; send the literal value "1".'
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        User API token or service token. Scopes (enforced server-side, not
        via OAuth): gateway:use, llm:invoke, admin, system.
    agentToken:
      type: http
      scheme: bearer
      description: >
        Per-AI-server agent secret (hashed lookup against agent_tokens),
        unrelated to user/service API tokens.
  requestBodies:
    OpaqueJSON:
      required: true
      content:
        application/json:
          schema:
            type: object
            description: >
              DTO defined by the Go handler (see api-surface.md and the
              gateway source); intentionally not duplicated here.
  responses:
    OpaqueJSON:
      description: >
        JSON DTO defined by the Go handler (see api-surface.md); errors use
        the Error envelope with a stable code.
      content:
        application/json:
          schema: { type: object }
...